Rename 알고리즘 스터디 - 큰 수의 법칙(그리디).cpp to 큰 수의 법칙(그리디).cpp
